gpc_reaction_set.h, gpc_reaction_set.cpp: Added more routines for the calculation of rate parameters and molar production rates.  Still not a full set, but sufficient for mopsc.
gpc_idealgas.cpp:
Correction to CalcGs function so that the correct number of terms are evalulated.

gpc_species.cpp:
Edit to Species::ThermoParams function so that it no longer throws exceptions.  Instead it just returns the most appropriate set of parameters.  In future it is suggested that this routine be able to send messages to the calling somehow to indicate that the parameters are not really valid.