swp_ensemble.h, swp_ensemble.cpp: Changed initialising constructor to require a mechanism argument.  Also made changes to operator= function to correctly copy the binary tree.
swp_treenode.h, swp_treenode.cpp:
Created initialising constructor which takes list of components and tracker variables as arguments.

swp_particledata.cpp:
Changes to operator= function to check if the components and tracker variables are the same for the LHS and RHS objects.  Uses the memcpy fucntion to quickly copy the components and trackers once the vectors are the correct size.