Discovered bug where whenever a copy of a PAH is made the combined sites for the copy is not updated correctly. This in turn affects the KMC-ARS jump process rates which are dependent on the number of combined site types.
Again, changes break pahtest3(-ens) and regression statistics had to be updated. Note that a slightly larger confidence interval (CI) width than the 99% CI width had to be used for both M0 and fv in order for the test to pass. Single run in the regression test may be an outlier.