Implemented fragmentation process for the spherical particle model solved using the stochastic weighted algorithm.
Solutions obtained using the uniform and size-dependent weighted symmetric kernels agree with the solutions obtained using DSA. For each fragmentation event, the particle mass is halved and its weight is doubled. Particle number and mass are conserved therefore it is not necessary to specify a weight transfer rule. For the uniform and size-dependent weighted erosion kernels either number (weightrule w1) or mass (weightrule w2) can be conserved but not both.