Can now specify the --wdotA4 flag for --flamepp
Users can now easily switch between postprocessing based on the
inception species concentration (default) or based on the molar rate of
production of the inception species by chemical reaction. This is only
relevant to the modelling of soot using the spherical and
PAH-PP/KMC-ARS models.

To-do: thresholds for inception, condensation and oxidation should be
inception species by default