Code to reproduce results in Preprint 159
Added a 5-member bay closure jump process which embeds a 5-member ring
and added a subsequent jump process which caps an embedded 5-member
ring which induces curvature. The code used to store the cartesian
coordinates of the edge carbon atoms, but now that PAHs can be curved,
all references to carbon atom positions are removed. The edge of each
PAH is instead described by a contiguous list of site types. Jump
processes which create bridges (phenyl addition and 6-member ring
oxidation at an armchair site by O2 and OH) are also commented out
because it is not possible to tell where the bridges are from a list of
site types alone. A database of certain combinations of site types is
stored where a jump process is not able to take place due to steric
hindrances. The rate constants of most of the jump processes are
ovehauled as they were either found to be wrong or lacked a proper
reference.