Removed database of site types for hindrances
It is not necessary to store such a database as such structures should
not arise in the first place. Take for example an armchair site
adjacent to a 6-member bay site. The previous commit would have
prevented 6-member ring growth from taking place on the armchair site,
thus resulting in structures with unrealistic arms. However, the 6-
member bay site should be allowed to close, converting the armchair
site into a 6-member bay site. This happens sufficiently fast such that
the above problem never takes place. Added analogous jump processes for
5-member rings such as 6-member ring growth on the free edge site of a
5-member ring.