Fixed random selection of primaries in aggregates
The way in which the SelectRandomSubparticleLoop function was
contructed was such that if the uniformly selected random integer
target was equal to the number of primaries in the left child, it would
never be selected. Surface reaction is performed by uniformly selecting
a primary and adjusting its composition. It was implemented incorrectly
and call to SelectRandomSubparticleLoop is instead made. Updated
pahtest1 regression test. Not as many tests were broken probably
because coagulation occurs frequently enough such that selection of
primaries is approximately random.