Track primary coordinates and do BCCA for coag
Track the coordinates of each primary particle in an aggregate for the
binary tree model (same has to be repeated for the PAH-PP/KMC-ARS
model). This feature can be switched on as follows in the sweep.xml
file:
    <particle ... track="primaryCoordinates">
    ...
    </particle>

Since the coordinates are tracked the primary particles in point
contact after a coagulation event are determined using Jullien's
ballistic cluster-cluster aggregation algorithm, which is more
realistic than uniformly selecting a primary particle in each
aggregate. I repeated the simulation conditions in his paper and was
able to reproduce his fractal dimension, Df = 2.06+-0.04 for d = 3 and
model II (without random impact). Little change in Df for model I
(random impact); did not attempt to reproduce Df as there are too many
uncertainties.

Newton bisection method code may have to be used to speed up
translation of particles. Resurrected radius of gyration feature.
Adjustment of particle radius and primary particle separation has to be
reexamined in light of the tracking of primary coordinates.