Solving MoMIC & Test class for unsteady Flamelets
A copy of the FlameLet class for Camflow solver has been created.
The purpose of this was to change the solve() function to solve chemistry to steady state and then use that as the IC for an unsteady flamelet calculation for soot.
This is how it is done in the CFD-Lagrangian Flamelet coupling. Here a test stop time is passed.

Source terms have been removed from cam_soot.cpp rateAll(). Introduce source terms one at a time (& recompile).
Currently, the nucleation terms are being solved using the restart(time) function. Test more later...