Added constantcalculated to Lewis choice
The option to use a constant, non-unity lewis number is available for any mechanism.
The new function calculates the lewis number for each species in an 1800K and 1 atm gas mixture using a binary diffusion coefficient from the Chapman Enskog Theory.
As it calculates each species, it prints the approriate line to terminal so you can copy and paste from the logfile into the needed LewisNumbersInput.xml to fixedFromFile input

An additional function was needed in sprog_transport_factory to be able to pass an ideal gas mixture when calculating Binary Diffusion coefficients. This function is a duplicate from the one that passes a regular mixture. This should not interfere with any current working MOPS versions.