dev-titania-phase: New melting point dependent phase transformation
The transformation is treated as a deferred process, similar to sintering. A primary particle freezes into the phase with the highest melting point. The melting point is given by T_melt = T_bulk(1 - A/d) for bulk melting temperature T_bulk and a parameter A (dimension: length). d is the primary particle size (spherical equivalent).
If the primary particle already contains solid phases, the liquid composition is transformed into an existing crystal phase with the phase determined probabilistically based on the relative composition of the crystal phases.

The feature is specified in sweep.xml as:

<transformation type="melting" name="Melting point transformation">
	<liquid id="Liq"/>	<!-- Specify the liquid phase -->
	<phase type="Anatase">	<!-- Define a phase change -->
		<A>1e-9</A> 			<!-- Parameter (m) -->
		<Tbulk>2075</Tbulk> 		<!-- Bulk melting temperature (K) -->
		<component id="Liq" dx="-1.0"/> <!-- Composition change -->
		<component id="An" dx="+1.0"/>
	</phase>
...
</transformation>