dev-titania-phase: implementation of the fixed crossover size and size-independent melting temperature model in the melting_model class
The melting model class is extended to support a fixed crossover size as an optional feature. The feature is enabled by specifying <crossover enable="true"/> and by supplying the min and max diameters (dmin and dmax) for each crystal phase. A size independent melting temperature is specified with A=0 and Tbulk=T, so that T_melt = T_bulk(1 - A/d).