dev-titania-videos: bintree particle tracking for videos
Tracks particle frame coordinates and orientation for producing videos. Primary particle coordinates, primary composition, frame coordinates and a frame orientation tracking vector is written to file at the end of each time step.

The feature is turned on in mops.inx within <output>
<trackparticle enable="true" pcount="1"/>

To do: output to binary and produce csv as a post-process.